Add license and re-arrange
When developing pylabeador, I studied the algorithm present in a C++
library to do syllable separation. This library was licensed with the
GPLv3-or-later. This means this works has to use a compatible license.

I am sharing this under the same license, and I am adding the original
copyright notice in one file only: pyleabeador/syllabify. This file
contains the core functionality of this library, in the syllabification
algorithms that I wrote following the C++ library.